+++ BeavR.xcodeproj/xcuserdata/pietro.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>LSE Now.xcscheme_^#shared#^_</key>
+		<key>BeavR.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
