+++ BeavR/Views/ConfirmEventSpotView.swift
