+++ BeavR/Views/HelpView.swift
