+++ BeavR/Views/NewEventView.swift
