+++ BeavRTests/BeavRTests.swift
@@ -1,14 +1,14 @@
 //
-//  LSE_NowTests.swift
-//  LSE NowTests
+//  BeavRTests.swift
+//  BeavRTests
 //
 //  Created by Pietro Canovari on 9/17/25.
 //
 
 import Testing
-@testable import LSE_Now
+@testable import BeavR
 
-struct LSE_NowTests {
+struct BeavRTests {
 
     @Test func example() async throws {
         // Write your test here and use APIs like `#expect(...)` to check expected conditions.
